camera/requirements.txt
django
requests
django-allauth
pillow
pypylon
matplotlib
djangorestframework
markdown
django-filter
django-currentuser
django-ckeditor
persiantools
